Make ouroboros-consensus warning free
Fixes #1659.

We now use the following warning flags for all components in consensus:

    -Wall
    -Wcompat
    -Wincomplete-uni-patterns
    -Wincomplete-record-updates
    -Wpartial-fields
    -Widentities
    -Wredundant-constraints
    -Wmissing-export-lists

This commit fixes all warnings produced by these flags.

+++ ouroboros-consensus-byron/src/Ouroboros/Consensus/Byron/Ledger/Conversions.hs
@@ -46,7 +46,7 @@ fromByronBlockNo :: CC.ChainDifficulty -> BlockNo
 fromByronBlockNo = coerce
 
 fromByronBlockCount :: CC.BlockCount -> SecurityParam
-fromByronBlockCount (CC.BlockCount k) = SecurityParam (fromIntegral k)
+fromByronBlockCount (CC.BlockCount k) = SecurityParam k
 
 fromByronEpochSlots :: CC.EpochSlots -> EpochSize
 fromByronEpochSlots (CC.EpochSlots n) = EpochSize n
@@ -59,7 +59,7 @@ toByronSlotNo :: SlotNo -> CC.SlotNumber
 toByronSlotNo = coerce
 
 toByronBlockCount :: SecurityParam -> CC.BlockCount
-toByronBlockCount (SecurityParam k) = CC.BlockCount (fromIntegral k)
+toByronBlockCount (SecurityParam k) = CC.BlockCount k
 
 {-------------------------------------------------------------------------------
   Extract info from genesis
